Characters way up there were being escaped as negative numbers...seems whacky.

+++ src/xmlwriter.cpp
@@ -67,7 +67,7 @@ std::string XmlWriter::escape( std::string sIn )
 		{
 			sOut += "&#";
 			char buf[4];
-			sprintf( buf, "%d", *i );
+			sprintf( buf, "%u", (unsigned int)*i );
 			sOut += buf;
 			sOut += ';';
 		}
